feat: set id for VIP-408
